+++ R/install_dev_deps.R
@@ -17,10 +17,11 @@
 #' + {testthat}
 #' + {rstudioapi}
 #'
-#' @param force_install If force_install is installed,
+#' @param force_install If force_install is TRUE,
 #'  then the user is not interactively asked
 #'  to install them.
 #' @param ... further arguments passed to the install function.
+#' @param dev_deps optional character vector of packages to install
 #'
 #' @export
 #'
@@ -31,6 +32,7 @@
 #'
 #' @return Used for side-effects
 install_dev_deps <- function(
+    dev_deps,
     force_install = FALSE,
     ...) {
   if (!force_install) {
@@ -60,6 +62,11 @@ install_dev_deps <- function(
     }
   }
 
+  if (missing(dev_deps)){
+    dev_deps <- getFromNamespace("dev_deps", "golem")
+  }
+  
+  
   for (
     pak in dev_deps
   ) {
@@ -69,45 +76,6 @@ install_dev_deps <- function(
   }
 }
 
-#' Install a single dev-dep
-#'
-#' Work the same way as [install_dev_deps()] when choosing installation
-#'   function.
-#'
-#' @param dev_dep a sing character giving the dependency to install
-#' @inheritParams install_dev_deps
-install_single_dev_dep <- function(
-  dev_dep = NULL,
-  force_install = FALSE,
-  ...) {
-    if (missing(dev_dep)) stop("Must provide arg to 'dev_dep'")
-    if (!is.character(dev_dep)) stop("Arg. 'dev_dep' must be character.")
-
-    # placeholder for 'temp_case'
-    tc <- character(0)
-    # go through cases as the install_dev_deps() function:
-    # case C0: force_install is FALSE and R is in non-interactive mode
-    if (isFALSE(force_install) && isFALSE(interactive())) tc <- "C0"
-    # case C1: force_install is FALSE and R is in interactive mode
-    if (isFALSE(force_install) && isTRUE(interactive())) tc <- "C1"
-    # case C2: force_install is TRUE and we have 'pak'
-    if (isTRUE(force_install) && isTRUE(rlang::is_installed("pak"))) tc <- "C2"
-    # case C3: force_install is TRUE and we do not have 'pak'
-    if (isTRUE(force_install) && isFALSE(rlang::is_installed("pak"))) tc <- "C3"
-
-    f <- switch(
-      tc,
-      C0 = "`install_dev_deps()` will not install dev dependencies in non-interactive mode if `force_install` is not set to `TRUE`.",
-      C1 = rlang::check_installed,
-      C2 = getFromNamespace("pkg_install", "pak"),
-      C3 = utils::install.packages
-    )
-    if (is.character(f)) {
-      warning(f)
-    } else {
-      f(dev_dep)
-    }
-  }
 
 dev_deps <- unique(
   c(
@@ -119,11 +87,12 @@ dev_deps <- unique(
     "dockerfiler",
     "fs",
     "here",
+    "httpuv",
     "pkgbuild",
     "pkgload",
     "processx",
-    "roxygen2",
     "renv",
+    "roxygen2",
     "rsconnect",
     "rstudioapi",
     "testthat",